Separated out forms into components so we can reuse headers etc. added some simple css.
