+++ frontend/src/App.css
@@ -3,3 +3,29 @@
   text-align: center;
 }
 
+.round-edge {
+  border: #9b9b9b;
+  border-radius: 1rem;
+}
+
+.primary-text-color {
+  color: rgb(39, 63, 27);
+}
+
+.primary-background-color {
+  background-color: rgb(237, 237, 237);
+}
+
+.std-padding {
+  padding: 0.5rem;
+  margin: 0.5rem 1rem;
+}
+
+.vertical-items {
+  display: flex;
+  flex-direction: column;
+}
+
+.hidden {
+  display: none;
+}
